APP/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.2
pandas>=2.1.4
numpy>=1.25.2
matplotlib>=3.8.2
streamlit>=1.29.0
pillow>=10.1.0
plotly>=5.17.0
joblib>=1.3.2
requests>=2.31.0
owslib>=0.29.0
